+++ lab5/functions.h
@@ -2,4 +2,4 @@
 
 #include <cmath>
 
-double f(double x, double y, double t);
+double f(double x, double y, int t);
